Refactor single PID indexer to separate component.
